# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swiss Life Triage Agent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/reference/priority": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Compute Priority",
        "operationId": "compute_priority",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriorityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriorityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/kb/documents": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Kb Documents",
        "operationId": "list_kb_documents",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "service_card",
                    "playbook",
                    "historical_ticket"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KbDocumentOut"
                  },
                  "title": "Response List Kb Documents"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kb/search": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Search Kb",
        "operationId": "search_kb",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KbSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Evidence"
                  },
                  "type": "array",
                  "title": "Response Search Kb"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assistant/ask": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Ask Assistant",
        "operationId": "ask_assistant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantAnswer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/export/submission": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Export Submission",
        "description": "Challenge-format records with our decisions filled in.\nUses the analyst-approved/edited decision when there is one, else the latest AI proposal.\nTODO: confirm the exact expected submission format with Swiss Life.",
        "operationId": "export_submission",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "challenge",
                "training",
                "manual",
                "email"
              ],
              "type": "string",
              "default": "challenge",
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Export Submission"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AssistantAnswer": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Citations"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "citations",
          "model"
        ],
        "title": "AssistantAnswer"
      },
      "AssistantRequest": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "ticket_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Id",
            "description": "Ground the answer in this ticket as well"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AssistantRequest"
      },
...
      "Evidence": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "service_card",
              "playbook",
              "historical_ticket"
            ],
            "title": "Kind"
          },
          "ref_id": {
            "type": "string",
            "title": "Ref Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "ref_id",
          "title",
          "snippet",
          "score"
        ],
        "title": "Evidence"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KbDocumentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "service_card",
              "playbook",
              "historical_ticket"
            ],
            "title": "Kind"
          },
          "ref_id": {
            "type": "string",
            "title": "Ref Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "has_embedding": {
            "type": "boolean",
            "title": "Has Embedding"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "ref_id",
          "title",
          "content",
          "meta",
          "has_embedding"
        ],
        "title": "KbDocumentOut"
      },
      "KbSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "K",
            "default": 5
          },
          "kinds": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "service_card",
                    "playbook",
                    "historical_ticket"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kinds"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "KbSearchRequest"
      },
...
      "PriorityRequest": {
        "properties": {
          "urgency": {
            "type": "string",
            "enum": [
              "Highest",
              "High",
              "Medium",
              "Low",
              "Lowest"
            ],
            "title": "Urgency"
          },
          "impact": {
            "type": "string",
            "enum": [
              "Highest",
              "High",
              "Medium",
              "Low",
              "Lowest"
            ],
            "title": "Impact"
          }
        },
        "type": "object",
        "required": [
          "urgency",
          "impact"
        ],
        "title": "PriorityRequest"
      },
      "PriorityResponse": {
        "properties": {
          "priority": {
            "type": "string",
            "enum": [
              "Highest",
              "High",
              "Medium",
              "Low",
              "Lowest"
            ],
            "title": "Priority"
          }
        },
        "type": "object",
        "required": [
          "priority"
        ],
        "title": "PriorityResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}